fix css, fix scan homepage

+++ app/models/skiresort.rb
@@ -71,7 +71,7 @@ class Skiresort < ActiveRecord::Base
 			end
 
 			if last && best == 0 	#there is an entry for today but the snowdate on the homepage was removed, so we delete the entry for the current day
-				last.destroy!
+				last.destroy! if last.time > Time.zone.now.beginning_of_day	
 			elsif last && best > 0 #there is an active entry for today so we update your entry
 				last.snow_height = best
 				last.save!
